fyp-extractive-qa/requirements.txt
docx2txt
pdf2image
pytesseract
Pillow
torch
transformers
fastapi